Create scopes for posts and categories and apply them through middleware
Create scope for ignoring private posts on client side, and a scope to
only display live categories, and a scope to only let published posts
to be displayed on the client side. The scopes are applied in
ClientScopes middleware where all the client routes will be
categorized under a group that has this middleware.